requirements.txt
firebase-admin==6.2.0
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3.0
gunicorn==21.2.0 